# evmrs: decouple driver and benchmarks from evmrs to be able to run other emvc interpreters