--- a/rust/src/lib.rs
+++ b/rust/src/lib.rs
@@ -36,18 +36,11 @@ compile_error!(
     Either disable it or enable one or all of `jumptable-dispatch`, `fn-ptr-conversion-expanded-dispatch` or `fn-ptr-conversion-inline-dispatch`."
 );
 
-#[cfg(not(feature = "custom-evmc"))]
-pub extern crate evmc_vm_tosca as evmc_vm;
-#[cfg(feature = "custom-evmc")]
-pub extern crate evmc_vm_tosca_refactor as evmc_vm;
-
 use llvm_profile_wrappers::{
     llvm_profile_enabled, llvm_profile_reset_counters, llvm_profile_set_filename,
     llvm_profile_write_file,
 };
-#[cfg(feature = "mock")]
-pub use types::MockExecutionContextTrait;
-pub use types::{u256, ExecutionContextTrait, MockExecutionMessage, Opcode};
+use types::u256;
 
 /// Dump coverage data when compiled with `RUSTFLAGS="-C instrument-coverage"`.
 /// Otherwise this is a no-op.