--- a/rust/src/types/mod.rs
+++ b/rust/src/types/mod.rs
@@ -3,10 +3,8 @@ mod amount;
 mod cache;
 mod code_analysis;
 mod code_reader;
-mod execution_context;
 pub mod hash_cache;
 mod memory;
-mod mock_execution_message;
 mod observer;
 #[cfg(feature = "needs-fn-ptr-conversion")]
 mod op_fn_data;
@@ -24,9 +22,7 @@ pub use cache::Cache;
 pub use cache::LocalKeyExt;
 pub use code_analysis::{AnalysisContainer, CodeAnalysis};
 pub use code_reader::{CodeReader, GetOpcodeError};
-pub use execution_context::*;
 pub use memory::Memory;
-pub use mock_execution_message::MockExecutionMessage;
 pub use observer::*;
 #[cfg(feature = "needs-fn-ptr-conversion")]
 pub use op_fn_data::OpFnData;