--- a/rust/src/types/op_fn_data.rs
+++ b/rust/src/types/op_fn_data.rs
@@ -4,9 +4,8 @@ use std::fmt::Debug;
 use crate::u256;
 use crate::{
     interpreter::{GenericJumptable, OpFn},
-    types::CodeByteType,
+    types::{CodeByteType, Opcode},
     utils::GetGenericStatic,
-    Opcode,
 };
 
 #[cfg(all(