--- a/rust/tests/ffi.rs
+++ b/rust/tests/ffi.rs
@@ -1,14 +1,15 @@
 #![allow(unused_crate_dependencies)]
+use common::{
+    evmc_vm::{Revision, StatusCode, StepStatusCode},
+    opcode::{ADD, PUSH0},
+    MockExecutionContextTrait, MockExecutionMessage,
+};
 #[cfg(not(feature = "custom-evmc"))]
 use driver::TX_CONTEXT_ZEROED;
 use driver::{
     get_tx_context_zeroed,
     host_interface::{self, null_ptr_host_interface},
     Instance, SteppableInstance, ZERO,
-};
-use evmrs::{
-    evmc_vm::{Revision, StatusCode, StepStatusCode},
-    MockExecutionContextTrait, MockExecutionMessage, Opcode,
 };
 
 #[test]
@@ -23,7 +24,7 @@ fn execute_can_be_called_with_mocked_context() {
         .return_const(TX_CONTEXT_ZEROED);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Push0 as u8];
+    let code = &[PUSH0];
     let result = instance.run(&host, &mut context, revision, &message, code);
     assert_eq!(result.status_code, StatusCode::EVMC_SUCCESS);
 }
@@ -35,7 +36,7 @@ fn execute_can_be_called_with_hardcoded_context() {
     host.get_tx_context = Some(get_tx_context_zeroed);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Push0 as u8];
+    let code = &[PUSH0];
     let result = instance.run_with_null_context(&host, revision, &message, code);
     assert_eq!(result.status_code, StatusCode::EVMC_SUCCESS);
 }
@@ -59,7 +60,7 @@ fn execute_handles_error_correctly() {
     host.get_tx_context = Some(get_tx_context_zeroed);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Add as u8]; // this will error because the stack is empty
+    let code = &[ADD]; // this will error because the stack is empty
     let result = instance.run_with_null_context(&host, revision, &message, code);
     assert_eq!(result.status_code, StatusCode::EVMC_STACK_UNDERFLOW);
 }
@@ -76,7 +77,7 @@ fn step_n_can_be_called_with_mocked_context() {
         .return_const(TX_CONTEXT_ZEROED);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Push0 as u8];
+    let code = &[PUSH0];
     let result = instance.run(
         &host,
         &mut context,
@@ -102,7 +103,7 @@ fn step_n_can_be_called_with_hardcoded_context() {
     host.get_tx_context = Some(get_tx_context_zeroed);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Push0 as u8];
+    let code = &[PUSH0];
     let result = instance.run_with_null_context(
         &host,
         revision,
@@ -152,7 +153,7 @@ fn step_n_handles_error_correctly() {
     host.get_tx_context = Some(get_tx_context_zeroed);
     let revision = Revision::EVMC_CANCUN;
     let message = MockExecutionMessage::default().to_evmc_message();
-    let code = &[Opcode::Add as u8]; // this will error because the stack is empty
+    let code = &[ADD]; // this will error because the stack is empty
     let result = instance.run_with_null_context(
         &host,
         revision,