This PR is the first step in decoupling the driver (and therefore also the rust benchmarks und fuzz tests) from evmrs. In the end the goal would be to be able to use those components also with other evmc interpreters (e.g. evmzero).

This PR moves all components of evmrs that are also needed by the driver, benchmarks and fuzz tests (reexport of evmc_vm, mocking infrastructure and opcode constants) into a new crate called common.
This also means that evmrs no longer exports any Rust symbols.

The driver still depends on evmrs, but only for linking to work out of the box.
Additionally the driver now exposes all features of evmrs so that the benchmarks and fuzz tests no longer need to depend on evmrs.

```
                   │ 2024-12-09T09:57#cf0a2ae#20-main/evmrs#performance │ 2024-12-11T09:24#d164688#20/evmrs#performance │
                   │                       sec/op                       │        sec/op          vs base                │
StaticOverhead/1/                                           1.770µ ± 1%             1.769µ ± 1%        ~ (p=0.846 n=20)
Inc/1/                                                      3.784µ ± 0%             3.883µ ± 1%   +2.62% (p=0.000 n=20)
Inc/10/                                                     3.784µ ± 0%             3.880µ ± 0%   +2.55% (p=0.000 n=20)
Fib/1/                                                      3.396µ ± 0%             3.472µ ± 0%   +2.24% (p=0.000 n=20)
Fib/5/                                                      12.01µ ± 0%             12.88µ ± 0%   +7.27% (p=0.000 n=20)
Fib/10/                                                     110.7µ ± 0%             123.9µ ± 0%  +11.88% (p=0.000 n=20)
Fib/15/                                                     1.089m ± 0%             1.193m ± 0%   +9.49% (p=0.000 n=20)
Fib/20/                                                     11.76m ± 0%             12.86m ± 0%   +9.38% (p=0.000 n=20)
Sha3/1/                                                     2.085µ ± 0%             2.102µ ± 1%   +0.82% (p=0.045 n=20)
Sha3/10/                                                    3.288µ ± 1%             3.300µ ± 0%   +0.38% (p=0.005 n=20)
Sha3/100/                                                   14.73µ ± 0%             14.54µ ± 0%   -1.30% (p=0.000 n=20)
Sha3/1000/                                                  150.7µ ± 0%             148.2µ ± 0%   -1.68% (p=0.000 n=20)
Arith/1/                                                    4.719µ ± 0%             4.780µ ± 1%   +1.29% (p=0.000 n=20)
Arith/10/                                                   8.582µ ± 1%             8.773µ ± 1%   +2.23% (p=0.000 n=20)
Arith/100/                                                  47.00µ ± 0%             47.96µ ± 0%   +2.03% (p=0.000 n=20)
Arith/280/                                                  152.4µ ± 0%             153.7µ ± 0%   +0.83% (p=0.000 n=20)
Memory/1/                                                   6.310µ ± 1%             6.318µ ± 0%        ~ (p=0.774 n=20)
Memory/10/                                                  10.86µ ± 1%             11.16µ ± 0%   +2.82% (p=0.000 n=20)
Memory/100/                                                 55.98µ ± 0%             56.93µ ± 0%   +1.69% (p=0.000 n=20)
Memory/1000/                                                499.6µ ± 0%             508.4µ ± 0%   +1.78% (p=0.000 n=20)
Memory/10000/                                               4.672m ± 0%             4.750m ± 0%   +1.66% (p=0.000 n=20)
Analysis/jumpdest/                                          1.765µ ± 1%             1.793µ ± 1%   +1.59% (p=0.001 n=20)
Analysis/stop/                                              1.771µ ± 1%             1.782µ ± 1%   +0.65% (p=0.013 n=20)
Analysis/push1/                                             1.769µ ± 1%             1.783µ ± 1%   +0.79% (p=0.001 n=20)
Analysis/push32/                                            1.775µ ± 1%             1.803µ ± 0%   +1.55% (p=0.000 n=20)
geomean                                                     20.43µ                  20.93µ        +2.45%
```